Use long doubles to store exp vectors used to compute markage.
This mitigates the issues with exp(t/tau) and exp(-t/tau) terms
over/underflowing when tau is too small with respect to the spike
times: on my system, if tau=1ms this happened for spike times larger
than a few hundred ms. Using long doubles, I can go up to about 10s.

Overall, there's still work to be done to endure that when the
exponential do blow up we raise an exception at the Python level, as
opposed to what's happening now (all distances being silently set to
zero).